Create dtype utils
